Improved launcher and scanner factory names and use build tags for journald in launcher instead of tailer
