+++ pkg/logs/input/windowsevent/launcher.go
@@ -20,8 +20,8 @@ type Launcher struct {
 	tailers          map[string]*Tailer
 }
 
-// New returns a new Launcher.
-func New(sources *config.LogSources, pipelineProvider pipeline.Provider) *Launcher {
+// NewLauncher returns a new Launcher.
+func NewLauncher(sources *config.LogSources, pipelineProvider pipeline.Provider) *Launcher {
 	return &Launcher{
 		sources:          sources,
 		pipelineProvider: pipelineProvider,
